CAMEL-21249 - Camel-Kamelets: Move Kamelets utils in Camel Kamelets component - Move DDB2 part to AWS DDB
