+++ components/camel-aws/camel-aws2-ddb/src/main/java/org/apache/camel/component/aws2/ddb/transform/serialization/gson/JavaTimeInstantTypeAdapter.java
@@ -15,7 +15,7 @@
  * limitations under the License.
  */
 
-package org.apache.camel.component.kamelet.utils.serialization.gson;
+package org.apache.camel.component.aws2.ddb.transform.serialization.gson;
 
 import java.lang.reflect.Type;
 import java.time.Instant;
