+++ components/camel-kamelet/pom.xml
@@ -85,12 +85,6 @@
             <artifactId>camel-mongodb</artifactId>
             <scope>provided</scope>
         </dependency>
-        <dependency>
-            <groupId>org.apache.camel</groupId>
-            <artifactId>camel-gson</artifactId>
-            <scope>provided</scope>
-        </dependency>
-
 
         <!-- TESTS  -->
         <dependency>
